requirements.txt
Django==3.2.12
djangorestframework==3.13.1
numpy==1.21.5
pandas==1.3.5
pandas_profiling==3.1.0
gunicorn==20.1.0
#azure-cli-ml
django-cors-headers==3.2.0
#mysqlclient
django-fine-uploader==0.2.0
celery==4.3.0
redis==3.2.0
eventlet
vine==1.3.0
azureml-sdk==1.40.0
azure-storage-blob==12.11.0
azureml-core==1.40.0
azure-devops
ruamel.yaml==0.17.21
sklearn
consoleprinter==93